Added images to introduction; Introduction library dimens added; SimpleSlide changed to FragmentSlide

+++ app/src/main/java/com/enoughspam/step/generalClasses/ScreenInfo.java
@@ -20,7 +20,7 @@ public class ScreenInfo {
         return mDisplayMetrics.widthPixels;
     }
 
-    public float getHeight() {
+    public int getHeight() {
         return mDisplayMetrics.heightPixels;
     }
